requirements.txt
pandas
sqlalchemy
psycopg2-binary
tabulate
